# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Online Course Platform API",
        "description": "API documentation for the online course backend built with Laravel.",
        "contact": {
            "email": "support@yourdomain.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Dev Server"
        }
    ],
    "paths": {
...
        "/api/v1/enroll": {
            "post": {
                "tags": [
                    "Enrollment"
                ],
                "summary": "Enroll in a course (via QR or enrollment code)",
                "operationId": "c9d691878d5480906025b0af52406f86",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "course_id"
                                ],
                                "properties": {
                                    "course_id": {
                                        "type": "integer"
                                    },
                                    "enrollment_code": {
                                        "type": "string"
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "enum": [
                                            "qrcode",
                                            "code"
                                        ]
                                    },
                                    "payment_reference": {
                                        "type": "string"
                                    },
                                    "verified": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Enrollment created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid code or unverified payment"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/enroll-from-payment/{paymentId}": {
            "post": {
                "tags": [
                    "Enrollment"
                ],
                "summary": "Enroll into multiple purchased courses after verified payment",
                "operationId": "ee3a08dc7c86dfeb86be0ea84af08c47",
                "parameters": [
                    {
                        "name": "paymentId",
                        "in": "path",
                        "description": "ID of the verified payment",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully enrolled into purchased courses"
                    },
                    "404": {
                        "description": "Payment not found or not verified"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/my-courses": {
            "get": {
                "tags": [
                    "Enrollment"
                ],
                "summary": "Get all courses the user is enrolled in",
                "operationId": "38d930678fb176287d4ea327ce355212",
                "responses": {
                    "200": {
                        "description": "List of enrolled courses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/lessons": {
            "post": {
                "tags": [
                    "Lessons"
                ],
                "summary": "Create a new lesson",
                "operationId": "199ffb4f15f3cfd703805ff876a746d4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "section_id",
                                    "title",
                                    "order"
                                ],
                                "properties": {
                                    "section_id": {
                                        "type": "integer"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    },
                                    "video_url": {
                                        "type": "string",
                                        "format": "url"
                                    },
                                    "order": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Lesson created successfully"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/lessons/{id}": {
            "put": {
                "tags": [
                    "Lessons"
                ],
                "summary": "Update an existing lesson",
                "operationId": "faf9781c137adb1e7a53392a5eb9ca85",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Lesson ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    },
                                    "video_url": {
                                        "type": "string",
                                        "format": "url"
                                    },
                                    "order": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lesson updated successfully"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Lesson not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Lessons"
                ],
                "summary": "Delete a lesson",
                "operationId": "9437b78d36fe3bbbbcb05d0b50bbcba5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Lesson ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lesson deleted successfully"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Lesson not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}